:sparkles: delete function operatiannal!

+++ backend/src/handlers/deleteRoutesFunctions.js
@@ -22,26 +22,29 @@ const deleteVideoById = (req, res) => {
   database
     .query("DELETE FROM video_category WHERE video_id = ?", [id])
     .then(([videoCategory]) => {
-      return videoCategory.affectedRows === 0
-        ? res.status(404).send("Not Found")
-        : res.status(204).send("video_category attachment deleted");
+      if (videoCategory.affectedRows !== 0) {
+        database
+          .query("DELETE FROM video WHERE id = ?", [id])
+          .then(([video]) => {
+            return video.affectedRows === 0
+              ? res.status(404).send("Not Found")
+              : res.sendStatus(204);
+          })
+          .catch((err) => {
+            console.error(err);
+            res.status(500).send("Error deleting a video");
+          });
+      } else {
+        console.log("youhou ça marche pas!")
+      }
+
     })
     .catch((err) => {
       console.error(err);
       res.status(500).send("Error deleting a video_category attachment");
     });
 
-  database
-    .query("DELETE FROM video WHERE id = ?", [id])
-    .then(([video]) => {
-      return video.affectedRows === 0
-        ? res.status(404).send("Not Found")
-        : res.sendStatus(204);
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error deleting a video");
-    });
+
 };
 
 const deleteCategoryById = (req, res) => {
